backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
langchain>=0.0.335
langchain-openai>=0.0.1
langchain-google-genai>=0.0.5
langchain-chroma>=0.0.1
sentence-transformers>=2.2.2
unstructured>=0.10.30
google-api-python-client>=2.108.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.4.2
python-multipart==0.0.9
requests>=2.31.0
beautifulsoup4==4.12.2
pandas>=2.1.1
numpy>=1.26.0
aiofiles==23.2.1
jinja2>=3.1.2
markdown>=3.5
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.2
faiss-cpu>=1.7.4
langchain-community[faiss]>=0.0.10
langchain-google-genai>=0.0.3
